try fix delete

+++ src/main/java/com/webapp/dao/impl/UserDaoImpl.java
@@ -2,9 +2,11 @@ package com.webapp.dao.impl;
 
 import com.webapp.dao.BaseDao;
 import com.webapp.dao.UserDao;
+import com.webapp.pojo.Job;
 import com.webapp.pojo.User;
 
 import java.sql.SQLException;
+import java.util.List;
 
 public class UserDaoImpl extends BaseDao implements UserDao {
     @Override
@@ -77,7 +79,20 @@ public class UserDaoImpl extends BaseDao implements UserDao {
 
     @Override
     public boolean delete(String emailId) {
-        String sql = "DELETE FROM `user` WHERE `email`=?";
+        User user = queryInfoByEmail(emailId);
+        if (user == null) {
+            return false;
+        }
+        int userId = user.getId();
+        String sql = "SELECT `job_id` FROM `post` WHERE user_id=?";
+        List<Job> jobs = queryForList(Job.class, sql, userId);
+        sql = "DELETE FROM `post` WHERE user_id=?";
+        update(sql, userId);
+        sql = "DELETE FROM `job` WHERE job_id=?";
+        for (Job job : jobs) {
+            update(sql, job.getId());
+        }
+        sql = "DELETE FROM `user` WHERE `email`=?";
         return update(sql, emailId) > 0;
     }
